+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v4.19.0
 ## 🚀 Feature
 - AndesMoneyAmountDiscount method used to expose contentDescription | Author: [@egeribon](https://github.com/egeribon)
+## 🛠 Fixes
+- Fix in the suffix text color for the `PERCENTAGE` variant in the `AndesAmountFieldSimple` component | Author: [@juanunzu](https://github.com/juanunzu)
 
 # v4.18.0
 ## 🚀 Feature